+++ pruebas/01-reading-list/Alber05/src/components/books/bookCard.module.scss
@@ -11,6 +11,10 @@
     overflow: hidden;
     filter: drop-shadow(1px 2px 3px #353838cb);
 
+    &:hover .synopsis {
+      right: 0;
+    }
+
     img {
       width: 100%;
       height: 302px;
@@ -84,6 +88,7 @@
       display: flex;
       align-items: center;
       gap: 0.5rem;
+      background-color: transparent;
       cursor: pointer;
       transition: 0.2s;
       font-family: "Amazon Ember Bold", sans-serif;
