+++ pruebas/01-reading-list/Alber05/src/components/emptyLibrary/emptyLibrary.module.css
@@ -6,17 +6,17 @@
   justify-content: space-around;
   align-items: center;
   min-height: calc(100vh - 250px);
-
-  @media (min-width: 768px) {
+}
+@media (min-width: 768px) {
+  .noBooksMain {
     margin: 170px auto 0 auto;
     min-height: calc(100vh - 170px);
   }
-
-  .noBooksTitle {
-    font-family: "Amazon Ember Heavy", sans-serif;
-  }
-
-  .noBooksImage {
-    image-rendering: optimizeQuality;
-  }
 }
+.noBooksMain .noBooksTitle {
+  font-family: "Amazon Ember Heavy", sans-serif;
+  color: #353838;
+}
+.noBooksMain .noBooksImage {
+  image-rendering: optimizeQuality;
+}/*# sourceMappingURL=noBooks.module.css.map */
