+++ pruebas/01-reading-list/Alber05/src/components/menu/Menu.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { NavLink } from "react-router-dom";
-import style from "./menu.module.css";
 import { PageSlider } from "./PageSlider";
-import CategoryFilter from "./CategoryFilter";
+import { CategoryFilter } from "./CategoryFilter";
+import { NavLinks } from "./NavLinks";
+import style from "./menu.module.css";
 
 export const Menu = () => {
   return (
@@ -12,24 +12,7 @@ export const Menu = () => {
           <PageSlider />
           <CategoryFilter />
         </div>
-        <div className={style.navLinksContainer}>
-          <NavLink
-            className={({ isActive }) =>
-              `${isActive ? `${style.link} ${style.active}` : style.link}`
-            }
-            to="/"
-          >
-            Todos los libros
-          </NavLink>
-          <NavLink
-            className={({ isActive }) =>
-              `${isActive ? `${style.link} ${style.active}` : style.link}`
-            }
-            to="/biblioteca"
-          >
-            Lista de lectura
-          </NavLink>
-        </div>
+        <NavLinks />
       </nav>
     </div>
   );
