+++ pruebas/01-reading-list/Alber05/src/components/menu/categoryFilter.module.css
@@ -1,0 +1,13 @@
+.categoryFilter .categoryFilterH4 {
+  color: #353838;
+}
+.categoryFilter .selectCategoryFilter {
+  border-radius: 1.2rem;
+  padding: 0.5rem;
+  border: none;
+  outline: none;
+  font-family: "Amazon Ember Bold", sans-serif;
+}
+.categoryFilter .selectCategoryFilter option {
+  padding: 1rem;
+}
