+++ pruebas/01-reading-list/Alber05/src/components/menu/menu.module.scss
@@ -18,40 +18,6 @@
       flex-direction: row;
     }
 
-    .navLinksContainer {
-      @media (min-width: 768px) {
-        width: 50%;
-        display: flex;
-        justify-content: flex-end;
-      }
-      .link {
-        color: #e3e5ec;
-        background-color: #353838;
-        padding: 1rem;
-        width: 15rem;
-        text-align: center;
-        filter: drop-shadow(1px 1px 2px #353838cb);
-
-        &:hover {
-          background-color: rgba(53, 56, 56, 0.7);
-        }
-
-        &:nth-child(1) {
-          border-radius: 1.2rem 0 0 1.2rem;
-        }
-        &:nth-child(2) {
-          border-radius: 0 1.2rem 1.2rem 0;
-        }
-      }
-
-      .link.active {
-        color: #f49400;
-        background-color: rgba(53, 56, 56, 0.8);
-        transform: scale(0.95);
-        font-family: "Amazon Ember Bold", sans-serif;
-      }
-    }
-
     .filtersContainer {
       display: flex;
       gap: 2rem;
@@ -60,18 +26,6 @@
         width: 50%;
         justify-content: flex-start;
         gap: 2rem;
-      }
-
-      .selectCategoryFilter {
-        border-radius: 1.2rem;
-        padding: 0.5rem;
-        border: none;
-        outline: none;
-        font-family: "Amazon Ember Bold", sans-serif;
-
-        option {
-          padding: 1rem;
-        }
       }
     }
   }
