+++ pruebas/01-reading-list/Alber05/src/components/menu/navlinks.module.css
@@ -1,0 +1,32 @@
+@media (min-width: 768px) {
+  .navLinksContainer {
+    width: 50%;
+    display: flex;
+    justify-content: flex-end;
+  }
+}
+.navLinksContainer .link {
+  color: #e3e5ec;
+  background-color: #353838;
+  padding: 1rem;
+  width: 15rem;
+  text-align: center;
+  filter: drop-shadow(1px 1px 2px rgba(53, 56, 56, 0.7960784314));
+  transition: background-color 0.2s ease-in;
+}
+.navLinksContainer .link:hover {
+  background-color: rgba(53, 56, 56, 0.6);
+}
+.navLinksContainer .link:nth-child(1) {
+  border-radius: 1.2rem 0 0 1.2rem;
+}
+.navLinksContainer .link:nth-child(2) {
+  border-radius: 0 1.2rem 1.2rem 0;
+}
+.navLinksContainer .link.active {
+  color: #f49400;
+  transform: scale(0.95);
+  font-family: "Amazon Ember Bold", sans-serif;
+  transition: background-color 1s;
+  background-color: rgba(53, 56, 56, 0.6);
+}
