+++ pruebas/01-reading-list/Alber05/src/components/menu/navlinks.module.scss
@@ -1,0 +1,35 @@
+.navLinksContainer {
+  @media (min-width: 768px) {
+    width: 50%;
+    display: flex;
+    justify-content: flex-end;
+  }
+  .link {
+    color: #e3e5ec;
+    background-color: #353838;
+    padding: 1rem;
+    width: 15rem;
+    text-align: center;
+    filter: drop-shadow(1px 1px 2px #353838cb);
+    transition: background-color 0.2s ease-in;
+
+    &:hover {
+      background-color: rgba(53, 56, 56, 0.6);
+    }
+
+    &:nth-child(1) {
+      border-radius: 1.2rem 0 0 1.2rem;
+    }
+    &:nth-child(2) {
+      border-radius: 0 1.2rem 1.2rem 0;
+    }
+  }
+
+  .link.active {
+    color: #f49400;
+    transform: scale(0.95);
+    font-family: "Amazon Ember Bold", sans-serif;
+    transition: background-color 1s;
+    background-color: rgba(53, 56, 56, 0.6);
+  }
+}
