+++ pruebas/01-reading-list/Alber05/src/components/menu/pageSlider.module.scss
@@ -1,0 +1,7 @@
+.pageSliderContainer {
+  .pageSliderh4 {
+  }
+
+  .sliderRange {
+  }
+}
