+++ pruebas/01-reading-list/Alber05/src/index.css
@@ -31,25 +31,21 @@ h1, h2, h3, h4, h5, h6 {
 }
 
 a {
-  font-weight: 500;
-  color: #646cff;
   text-decoration: none;
+  color: #353838;
 }
-a:hover {
-  color: #535bf2;
-}
+
 
 html{
   font-size: 62.5%;
 }
 
 body {
-
   font-family: "Amazon Ember Light", sans-serif;
   font-size: 1.6rem;
   background: #E3E5EC;
   background-repeat: no-repeat;
-  color: black;
+
 }
